Use Serilog for file and console logging (#6)
Introduce Serilog + nest appsettings.json files

Now Serilog is used to log into files and on the console to reflect a
real scenario.
Just for a better overview, stage specific appsettings.<stage>.json
files are now nested on the original appsettings.json to make it obvious
in which order they are applied.