Fix by SonarQube (re)marked issues (#7)
Now the Serilog logger is closed and flushed asynchronously at the end,
because the host instance is already run async and so the whole
top-level statement becomes async automatically. This change addresses
the by SonarQube remarked issue csharpsquid:S6966

Removed a string interpolation from a log message, which should fix
SonarQube issue csharpsquid:S2629 and in theory Roslyn issues
roslyn:CA2254 and external_roslyn:CA2254 as well.

Due the CancellationTokenSource implements IDisposable and this is used
within the CancellationTokenFactory, the latter has to implement
IDisposable too to ensure that the hold CancellationTokenSource is
freed up properly. This addresses SonarQube issue csharpsquid:S2930.