Simplify cancellation token creation and handling (#8)
Since a CancellationTokenSource was needed that provides a possibility
to create an event handler to register at the Console.CancelKeyPress
event, the not really factory like CancellationTokenFactory class was
renamed into ConsoleCancellationTokenSource and inherited instead from
CancellationTokenSource. By then duplicated and obsolete code was
removed and the already existing event handler creation method was
renamed.
So, now this class purpose and behavior should be more clear.

Extended the team-shared dictionary by 'hostbuilder' entry.