Add GitVersion to build w/ GitFlow configuration (#9)
Integrated GitVersion GitHub actions for automated artifact versioning
based on _SemVer2_.
To do so, it was necessary to use the GitVersion `command` step instead
of the `execution` step, because not all needed functionality was
available.
Because the GitVersion action searches for `*.yml` files only, but we
use `*.yaml`, it was necessary to define specifically GitVersion.yaml as
config file.
Further, it does not update any projects' version, so parameter
`/updateprojectfiles` must be defined.
Finally, to ensure that all by GitVersion generated variables and values
are available, even when we use the `command` step, the output type
`buildserver` is used to get those written into `$GITHUB_ENV`. So all
variables can be used later on as generally available environment
variables.