Fix NU1603 Castle.Core 4.0.0 caused by NSubstitute
The build warning NU1603 was caused by using AutoFixture.AutoNSubstitute
that has a reference of NSubstitute v2.0.3 that references Castle.Core
v4.0.0, but on dotnet restore it's tried to resolve as v4.0.1 and ends
up with 4.1.0.
To get rid of that behavior, that as well slows down the whole restore
process dramatically, NSubstitute was installed as direct dependency of
the test project.