Add tests + code coverage w/ build integration (#10)
Add tests + code coverage w/ build integration

In general a test project was added and tried to cover possible and
realistic scenarios. Furthermore, now during the test execution a code
coverage report is created. The coverage information was integrated with
the SonarQube analysis too.
Both reports, test and coverage, are published as build artifacts.
Furthermore, the GitHub Actions logger was added and slightly configured
to provide more test insides within the build summary.

Please note as the App class itself does not really do anything, there's
only one test that let the loop run and simulate a graceful
cancellation.

Beside of testing the dependency registration, the registered lifetime
of classes is tested too to highlight any changes on that by failing
tests.

After using package `AutoFixture.AutoNSubstitute` the build warning
NU1603 occurred, because of a reference to `NSubstitute v2.0.3` that
references `Castle.Core v4.0.0`, but on `dotnet restore` it was tried to
resolve as v4.0.1 and ended up with 4.1.0.
To get rid of that behavior, that as well slows down the whole restore
process dramatically, NSubstitute was installed as direct dependency of
the test project.