Add bug report issue template (#29)
Add bug report issue template as Markdown

Added an issue template that represents a bug report to make it easier,
more convenient and efficient to create reports each time with the same
structure. In addition, this should assist users to provide needed
information for us to understand and reproduce the bug.

The decision was made to create this template as Markdown template for
being able to gain some experience with this kind of issue template too.
(The user story uses the YAML form approach already.)

Related item: #25