Add DurationTimeParser to parse time formats
Implemented a DurationTimeParser that is capable to parse different time
formats in different variations like duration format
"[d.]HH:mm:ss[.fff]" and a human-readable duration format with units
like "7s", "7sec", "1second" or "7seconds". For the later a space
between vale and unit is possible too.
In addition, this parse does an unrolling and rolling over of values
into the next bigger or smaller unit, meaning 69.5 seconds will result
in "00:01:09.500".

Supported human-readable units are:
- Millisecond: ms, millisecond, milliseconds
- Second: s, sec, second, seconds
- Minute: m, min, minute, minutes
- Hour: h, hour, hours
- Day: d, day, days

In addition, unit tests were added as well to ensure the parser works
correctly.

Item: #21