Modify logging configuration to avoid Serilog bug
Serilog is not able to merge the "WriteTo" section properly, it just
replaces the section with the section of the latest added
appsettings.json file. In our case the development environment specific
configuration during local development. So, that we ended up with two
console configurations, instead of each one file- and console
configuration.
This was caused by the fact that the general appsettings.json file
contained a file- and console configuration (in this order) and the
development environment specific one only console configuration.
This issue was simply workaround by changing the order so that the
console configuration comes first. Because of that change, console is in
both files on first position and its configuration is correctly
overwritten.

Item: #21