Introduced default/fallback Serilog configuration
From now Serilog is configured in a minimal way to log to the console
with a very simple message template, when no configuration was defined
within the appsettings.json file or it does not exist in general.
As this was the last non-default configured part of the application
contained in the appsettings.json file, the entire file has now become
optional.

Item: #21