Add configuration scope handling
Introduced support for scoped configuration management. These scopes are
system-level, user-level (global) and the current working directory
(local).
In addition the configuration scopes are validated and handled by the
`config set` command.

In regards of the configuration handling, all configuration file path
related code was extracted in scope specific classes to align with
single responsibility- and separation of concerns principles.

Enhanced testing coverage with comprehensive tests for scoped configurations.
For being able to test internal methods, the `InternalsVisibleTo`
property was added to the *.csproj file.

Item: #21