Introduced EnvironmentAbstractions for testability
For being able to test the configuration file path logic of available
configuration level (system, global and local) it was necessary to
decouple the .NET Environment implementation. This was achieved with
EnvironmentAbstractions.
This change needed much more modification in existing tests to realign
them.

Furthermore, the test coverage in general was a lot increased to cover
most possible scenarios.

Updated existing services to use `Scoped` lifetime as this makes more
sense for this applications purpose and to reduce object creation that
is not needed.

Fixed minor typos and updated XML documentation accordingly.

Item: #21