Fix services lifetime issue
With scoped services it was not possible to run the application in
debug mode. Some services were not resolvable. At this point in time it
is unclear if this is an issue related to Rider or the code itself,
hence it was rolled back to transient lifetime for all services.

Item: #21